Add --stl=<name> option to make-standalone-toolchain.sh
This modifies make-standalone-toolchain.sh to support --stl=stlport,
which allows one to build a standalone NDK toolchain that uses STLport
instead of the GNU libstdc++ library.

This allows building the ASan runtime for Android using the Chromium
NDK checkout directly.

Tested on Linux by invoking:

  build/tools/make-standalone-toolchain.sh \
    --stl=stlport \
    --install-dir=/tmp/ndk-arm-stlport \
    --system=linux-x86_64

This is a back-port of the following AOSP NDK patch:
https://android-review.googlesource.com/#/c/54520/

BUG=NONE